requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cryptography==35.0.0
Deprecated==1.2.13
docker==5.0.3
english==2020.7.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-restx==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
monotonic==1.6
neo4j==4.1.0
neotime==1.7.4
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-exporter-jaeger-proto-grpc==1.6.2
opentelemetry-exporter-jaeger-thrift==1.6.2
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-flask==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
opentelemetry-util-http==0.26b1
packaging==21.2
pansi==2020.7.3
prompt-toolkit==2.0.10
protobuf==3.19.1
py2neo==2021.0.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dotenv==0.19.1
python-json-logger==0.1.11
pytz==2020.1
requests==2.26.0
six==1.15.0
thrift==0.15.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.13.3
zipp==3.1.0
